habit_tracker.py
class HabitTracker:
    def __init__(self):
        """Initialize a new HabitTracker with an empty list of habits."""
        self.habits = []

    def add_habit(self, habit):
        """Add a new habit to the tracker."""
        self.habits.append(habit)

    def delete_habit(self, habit_name):
        """Delete a habit by its name."""
        self.habits = [habit for habit in self.habits if habit.get_name != habit_name]

    def get_habit(self, habit_name):
        """Retrieve a habit by its name."""
        for habit in self.habits:
            if habit.get_name == habit_name:
                return habit
        return None

    def analyze_habits(self):
        """Placeholder for habit analysis functionality."""
        # This method will later use functions from the 'analyse.py' module.
        pass




